Finilizing v2.0
 - Fixed pep8 issues
 - Added documentation
 - Changed package structure
