--- docs/topics/admin_integration.rst
@@ -1,6 +1,0 @@
-
-Admin Integration
-=================
-
-Prior to version 1.4 ``django_measurement.admin.MeasurementAdmin`` class was needed.
-This class is no longer needed, deprecated and will be removed in version 1.5.
