+++ pytest.ini
@@ -2,7 +2,6 @@
 norecursedirs=venv docs
 DJANGO_SETTINGS_MODULE=tests.settings
 addopts = --tb=short --pep8 --flakes -rxs
-pep8maxlinelength=99
 pep8ignore=
     setup.py ALL
     runtests.py ALL
@@ -10,5 +9,3 @@ pep8ignore=
 flakes-ignore=
     setup.py ALL
     runtests.py ALL
-    django_measurement/measure.py ImportStarUsed
-    django_measurement/utils.py UnusedImport
