+++ tests/models.py
@@ -1,40 +1,40 @@
 from django.db import models
-from django_measurement import fields
+from django_measurement.models import MeasurementField
 from measurement import measures
 
 
 class MeasurementTestModel(models.Model):
-    measurement_distance = fields.MeasurementField(
+    measurement_distance = MeasurementField(
         measurement=measures.Distance,
         max_value=3.0, min_value=1.0,
         blank=True, null=True,
     )
-    measurement_distance_km = fields.MeasurementField(
+    measurement_distance_km = MeasurementField(
         measurement=measures.Distance,
         unit_choices=(('km', 'km'),),
         max_value=3.0, min_value=1.0,
         blank=True, null=True,
     )
 
-    measurement_weight = fields.MeasurementField(
+    measurement_weight = MeasurementField(
         measurement=measures.Weight,
         max_value=3.0, min_value=1.0,
         blank=True, null=True,
     )
 
-    measurement_speed = fields.MeasurementField(
+    measurement_speed = MeasurementField(
         measurement=measures.Speed,
         max_value=3.0, min_value=1.0,
         blank=True, null=True,
     )
 
-    measurement_temperature = fields.MeasurementField(
+    measurement_temperature = MeasurementField(
         measurement=measures.Temperature,
         max_value=3.0, min_value=1.0,
         blank=True, null=True,
     )
 
-    measurement_speed_mph = fields.MeasurementField(
+    measurement_speed_mph = MeasurementField(
         measurement=measures.Speed,
         unit_choices=(('mi__hr', 'mph'),),
         max_value=3.0, min_value=1.0,
